Extract awards.js from gameLogic.js (closes #122)
