+++ src/screens/ChroniclesScreen.jsx
@@ -2,7 +2,8 @@ import { useState, useEffect } from 'react'
 import RoundChat from '../components/RoundChat.jsx'
 import CombatantSheet from '../components/CombatantSheet.jsx'
 import { btn, tab } from '../styles.js'
-import { tallyReactions, groupRoomsForHistory, computeSeriesStandings, computeSeriesAutoAwards, computeSeasonToneDisplay, AWARD_TYPE_LABELS } from '../gameLogic.js'
+import { tallyReactions, groupRoomsForHistory, computeSeriesStandings, computeSeasonToneDisplay } from '../gameLogic.js'
+import { computeSeriesAutoAwards, AWARD_TYPE_LABELS } from '../awards.js'
 import { slist, getAwardsForScope, createAutoAwards } from '../supabase.js'
 import { downloadFile, formatRoomAsText, formatSeriesAsText } from '../export.js'
